Bug Fix: Fixed wrong class types in TableFormats (BugId:  969, 971, 972, 979)  - Updated TableFormatTest

+++ src/main/java/net/nikr/eve/jeveasset/data/sde/Item.java
@@ -141,7 +141,7 @@ public class Item implements Comparable<Item>, ItemType {
 		return name;
 	}
 
-	public long getPriceBase() {
+	public double getPriceBase() {
 		return price;
 	}
